Added Task ownership. Authorization is completed.

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './tasks.entity';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskStatus } from './task.status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,12 +14,20 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) {}
     
-    getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User,
+        ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const task = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User
+        ): Promise<Task> {
+        const task = await this.taskRepository.findOne({
+            where: { id, userId: user.id }
+        });
 
         if(!task){
             throw new NotFoundException(`Task wiht Id: ${id} not found`);
@@ -28,21 +37,24 @@ export class TasksService {
     }
 
 
-    async deleteTaskById(id: number): Promise<void> {
-        const task = await this.taskRepository.delete(id);
+    async deleteTaskById(id: number, user: User): Promise<void> {
+        const task = await this.taskRepository.delete({ id, userId: user.id });
 
         if(task.affected === 0) {
             throw new NotFoundException(`Task with Id: ${id} not Found`);
         }
     }
 
-    async updateTaskById(id: number, status: TaskStatus): Promise<void> {
-        const task = await this.getTaskById(id);
+    async updateTaskById(id: number, status: TaskStatus, user: User): Promise<void> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         task.save()
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User,
+        ): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 }
